… benchmark/mmmu/bench_hf.py
… python/sglang/srt/managers/mm_utils.py
… python/sglang/srt/managers/multimodal_processors/minicpm.py
… python/sglang/srt/managers/multimodal_processors/qwen_vl.py
… python/sglang/srt/managers/schedule_batch.py
… python/sglang/srt/model_executor/model_runner.py
… python/sglang/srt/models/clip.py
… python/sglang/srt/models/deepseek_janus_pro.py
… python/sglang/srt/models/deepseek_vl2.py
… python/sglang/srt/models/gemma3_mm.py
… python/sglang/srt/models/llava.py
… python/sglang/srt/models/llavavid.py
… python/sglang/srt/models/minicpmo.py
… python/sglang/srt/models/minicpmv.py
… python/sglang/srt/models/mllama.py
… python/sglang/srt/models/qwen2.py
… python/sglang/srt/models/qwen2_5_vl.py
… python/sglang/srt/models/qwen2_vl.py
… python/sglang/srt/openai_api/adapter.py
… python/sglang/srt/utils.py
… test/srt/test_vision_openai_server.py
… test/srt/test_vlm_accuracy.py