<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Due to the loose connection between multimodal data and their pad values in the prior design, some vlms does not support radix-cache and chunked-prefill.

More specifically, for a batched request, or sometimes mixed batch, especially when `radix-cache` is enabled or audio is involved, some multimodal tokens from input_ids are truncated, making it difficult to judge the boundary of multimodal data, and also which data is actually being prefixed, or which data needs to be embedded.


<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

<!-- Describe the changes made in this PR. -->

1. A strong connection between multimodal data and its pad_value is be managed by `MultimodalDataItem`
  * this way, even if the batch is mixed/truncated, we can always find the right data to embed and scatter
After this modification:

2. Radix cache can be enabled for most vlms, and the data needed to be embedded is restricted to those does exist in input_ids
## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.