@mickqian 2025-03-26
mmmu results:

before:

|     |  accuracy  |  time elapsed  |
| --- | --- | --- |
|   Gemma-3-4b-it |    0.409  |  202.3 |
|   Qwen2.5-VL-7B-Instruct |    0.477   |  230 |
|   Minicpmv |    0.423   |  256 |

now:

|     |  accuracy  |  time elapsed  |  speedup (%) |
| --- | --- | --- | --- |
|   Gemma-3-4b-it |    0.411   |  166.6 |  21.4 |
|   Qwen2.5-VL-7B-Instruct |    0.477   | 170  | 35.3  |
|   Minicpmv |    0.423   |  202 |   26.7 |