@Swipe4057 2025-03-27
@mickqian
Hello, has this bug with long generations in Gemma 3 been fixed? I checked the current sglang 0.4.4.post2 and I'm getting the same error.
https://github.com/sgl-project/sglang/pull/4424#issuecomment-2737301031